+++ src/Decoder.ts
@@ -80,6 +80,7 @@ export interface Decoder<T> {
  *   DecoderType<typeof truthy>      // boolean
  *
  */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type DecoderType<D extends Decoder<any>> = D extends Decoder<infer T> ? T : never;
 
 function noThrow<T, V>(fn: (value: T) => V): (blob: T) => DecodeResult<V> {